FaceDetectionAPI/requirements.txt
fastapi
opencv-python
numpy
Pillow